Fix line number being reset on each Indented write_str call
In most realistic scenarios, the `write_str` method provided by
core::fmt::Write implementations is called several times, as this is
what allows to implement allocation-free string interpolation. When this
happens, each string passed to `write_str` can only be assumed to be a
slice of the text to be written.

However, the `write_str` implementation provided by the `Indented`
struct does not keep track of the number of lines written so far,
relying on the input string slice to represent the line count of the
whole text. As a consequence, when several strings are written out,
formatters will be provided with incorrect line count information:

```
indented.write_str("hello")?; // ind = 0 => Format line_no = 0
indented.write_char('\n')?;
indented.write_str("world")?; // ind = 0, but this is not the first line!
```

To improve on this, let's make Indented remember the number of lines
that have been written so far, and pass that to the formatters.

While at it, I've added the `empty_lines` test as a sanity check for
proper empty line handling, which should be the same as before. I've
also added the `several_interpolations_keep_monotonic_line_numbers`
test to check that these changes work as expected. Previously, that test
failed with this output:

```
thread 'tests::several_interpolations_keep_monotonic_line_numbers' panicked at 'assertion failed: `(left == right)`
  left: `"verify\n this\n  and this"`,
 right: `"verify\nthis\nand this"`'
```